requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
openapi-codec==1.3.2
optional-django==0.3.0
pandas==1.3.4
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
react==4.3.0
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.6
